backend/requirements.txt
Django==2.2.20
pytz==2022.1
sqlparse==0.4.2
djangorestframework==3.12.4
django-extra-fields==3.0.2
Pillow==9.1.1
psycopg2-binary==2.9.3